pyproject.toml
[tool.poetry]
name = "notification-discord-bot"
version = "0.1.0"
description = ""
authors = ["seendsouza <0xseen64@gmail.com>", "nazariyv <naz@renft.io>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
"discord.py" = "1.7.3"
python-dotenv = "0.20"
requests = "^2.28.1"
tweepy = "^4.10.1"

[tool.poetry.dev-dependencies]
black = "^22.6.0"
mypy = "^0.971"
pylint = "^2.14.5"
types-requests = "^2.28.5"
pytest = "^7.1.3"

[[tool.mypy.overrides]]
module = ["discord", "tweepy"]
ignore_missing_imports = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
